+++ cmd/unmount_all.go
@@ -34,8 +34,9 @@
 package cmd
 
 import (
-	"blobfuse2/common"
 	"fmt"
+
+	"github.com/Azure/azure-storage-fuse/v2/common"
 
 	"github.com/spf13/cobra"
 )