+++ internal/base_component.go
@@ -34,10 +34,11 @@
 package internal
 
 import (
-	"blobfuse2/common"
-	"blobfuse2/internal/handlemap"
 	"context"
 	"syscall"
+
+	"github.com/Azure/azure-storage-fuse/v2/common"
+	"github.com/Azure/azure-storage-fuse/v2/internal/handlemap"
 )
 
 // BaseComponent : Base implementation of the component interface
@@ -45,8 +46,6 @@ type BaseComponent struct {
 	compName string
 	next     Component
 }
-
-const name = "BaseComponent"
 
 var _ Component = &BaseComponent{}
 
@@ -62,7 +61,7 @@ func (base *BaseComponent) SetName(name string) {
 	base.compName = name
 }
 
-func (base *BaseComponent) Configure() error {
+func (base *BaseComponent) Configure(isParent bool) error {
 	return nil
 }
 
@@ -319,7 +318,7 @@ func (base *BaseComponent) InvalidateObject(name string) {
 
 func (base *BaseComponent) FileUsed(name string) error {
 	if base.next != nil {
-		base.next.FileUsed(name)
+		return base.next.FileUsed(name)
 	}
 	return nil
 }