+++ internal/component_options.go
@@ -34,8 +34,9 @@
 package internal
 
 import (
-	"blobfuse2/internal/handlemap"
 	"os"
+
+	"github.com/Azure/azure-storage-fuse/v2/internal/handlemap"
 )
 
 type CreateDirOptions struct {