+++ main.go
@@ -34,13 +34,13 @@
 package main
 
 import (
-	"blobfuse2/cmd"
-	_ "blobfuse2/common/log"
+	"github.com/Azure/azure-storage-fuse/v2/cmd"
+	_ "github.com/Azure/azure-storage-fuse/v2/common/log"
 )
 
 //go:generate ./cmd/componentGenerator.sh $NAME
 //  To use go:generate run command   "NAME="component" go generate"
 
 func main() {
-	cmd.Execute()
+	_ = cmd.Execute()
 }